Moved sampling of initial points to its own file
